+++ src/content-helper/editor-sidebar/performance-stats/component-panel-referrers.tsx
