+++ src/content-helper/editor-sidebar/performance-stats/model.ts
