+++ src/content-helper/editor-sidebar/performance-stats/performance-stats.scss
